+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Get upgrade notes from Sprockets 3.x to 4.x at https://github.com/rails/sprocket
 - Allow assets already fingerprinted to be served through `Sprockets::Server`
 - Do not fingerprint files that already contain a valid digest in their name
 - Remove remaining support for Ruby < 2.4.[#672](https://github.com/rails/sprockets/pull/672)
+- Embed sourcesContent in source maps. [#697](https://github.com/rails/sprockets/pull/697)
 
 ## 4.0.2